requirements.txt
fastapi
uvicorn
pandas
psycopg-binary
google-generativeai
python-dotenv
langgraph
langchain-google-genai
langchain-openai
openpyxl
pydantic
typing_extensions
langchain-google-genai